# Faster pre-processing for gray image input