<!--
Thank you for submitting a YOLOv5 🚀 Pull Request! We want to make contributing to YOLOv5 as easy and transparent as possible. A few tips to get you started:

- Search existing YOLOv5 [PRs](https://github.com/ultralytics/yolov5/pull) to see if a similar PR already exists.
- Link this PR to a YOLOv5 [issue](https://github.com/ultralytics/yolov5/issues) to help us understand what bug fix or feature is being implemented.
- Provide before and after profiling/inference/training results to help us quantify the improvement your PR provides (if applicable).

Please see our ✅ [Contributing Guide](https://github.com/ultralytics/yolov5/blob/master/CONTRIBUTING.md) for more details.
-->

Using opencv cvtColor function is significantly faster than np.tile method to enforce 3 channel input


```
%%timeit -r 7 -n 100
im_cv_3ch=cv.cvtColor(im,cv.COLOR_GRAY2BGR)
13.2 ms ± 250 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

%%timeit -r 7 -n 100
im_np_3ch=np.tile(im[..., None], 3)
119 ms ± 2.62 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

(Test done on Windows machine with Intel i5 CPU)