@glenn-jocher 2022-08-18
@cher-liang 10x speedup, very nice!

PR is merged. Thank you for your contributions to YOLOv5 🚀 and Vision AI ⭐